Split up the deployment module.
